chore: cleanup arm64 disassembler related code
